refactor: refactored product handler and application routes logic to make product follow the same structure as the other entities.
